+++ CHANGELOG.md
@@ -1,9 +1,13 @@
-# master
+# 1.0.0-rc.1
 
 - Introduce animations alongside moving [#13](https://github.com/muukii/Rideau/pull/13)
 - Fix scrolling-lock
 - Fix getting content inset from UIScrollView
 - Set identifier to NSLayoutConstraint that added inside Rideau
+
+# 1.0.0-beta.4
+
+- Swift5
 
 # 1.0.0-beta.3